app1/pyproject.toml
[project]
name = "app1"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Yang Qin", email = "qin.yang@emotion-tech.co.jp" }
]
dependencies = [
    "fastapi>=0.109.2",
    "uvicorn>=0.27.0.post1",
    "python-dotenv>=1.0.1",
    "pydantic>=2.6.1",
]
readme = "README.md"
requires-python = " >= 3.11"


[project.scripts]
hello = "app1:hello"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/app1"]
